+++ software/frequency_relay/frequency_relay.c
@@ -46,61 +46,6 @@ timingLog_t timing_log = {0};
 freqData_t freq_data;
 system_status_t system_status;
 
-void fau_isr(void* context, alt_u32 id) {
-	// placeholder for return status of if a higher priority task
-	// was blocked from a resource liberated in this ISR
-	BaseType_t xHigherPriorityTask = pdFALSE;
-
-	// read 32 bit N counter from FAU
-	freq_data.n = IORD_ALTERA_AVALON_PIO_DATA(FREQUENCY_ANALYSER_BASE);
-
-	// give the semaphore (aka signal to the task that data is available)
-	// also if there is a higher priority task waiting on that resource
-	// set xHigherPriorityTask high
-	xSemaphoreGiveFromISR(peakReadySem, &xHigherPriorityTask);
-
-	// if there is a higher priority task, switch to it before resuming
-	portEND_SWITCHING_ISR(xHigherPriorityTask);
-
-}
-
-void button_isr(void* context, alt_u32 id) {
-	// placeholder for return status of if a higher priority task
-	// was blocked from a resource liberated in this ISR
-	BaseType_t xHigherPriorityTask = pdFALSE;
-
-	int button_input = IORD_ALTERA_AVALON_PIO_DATA(PUSH_BUTTON_BASE);
-
-	// clear the hardware interrupt
-	IOWR_ALTERA_AVALON_PIO_EDGE_CAP(PUSH_BUTTON_BASE, 0xF);
-
-	// add data to mailbox, if higher task blocked cause queue
-	// make xHigherPriorityTask = pdTRUE
-	xQueueSendFromISR(buttonCmdQ, &button_input, &xHigherPriorityTask);
-
-	// if there is a higher priority task, switch to it before resuming
-	portEND_SWITCHING_ISR(xHigherPriorityTask);
-}
-
-void kbd_isr(void* context, alt_u32 id) {
-	(void)context;
-	(void)id;
-	BaseType_t xHigherPriorityTask = pdFALSE;
-
-	// read the full 32bit register
-	uint32_t ps2_reg = IORD_ALT_UP_PS2_PORT_DATA_REG(PS2_BASE);
-
-	// bit 15 is RVALID (if data valid)
-	if (ps2_reg & 0x8000) {
-		// extract only the 8bit key code (0-7)
-		int key_code = ps2_reg & 0xFF;
-
-		xQueueSendFromISR(kbdQ, &key_code, &xHigherPriorityTask);
-	}
-
-	portEND_SWITCHING_ISR(xHigherPriorityTask);
-}
-
 void debug_consumer_task(void *pvParameters) {
 	// TODO: remove once tested and working correctly
 	int button_rx;
